api/requirements/base.txt
# ------------------------------------------------------------------------------
# DJANGO
# ------------------------------------------------------------------------------
Django==2.2.7
django-environ==0.4.5
django-cors-headers==3.1.1
djangorestframework==3.10.3
django-simple-captcha==0.5.12


# ------------------------------------------------------------------------------
# POSTGRES
# ------------------------------------------------------------------------------
psycopg2-binary==2.8.4

# ------------------------------------------------------------------------------
# WSGI SERVER
# ------------------------------------------------------------------------------
gunicorn==19.9.0
